requirements.txt
autopep8==1.4.4
Click==7.0
feedparser==5.2.1
Flask==1.0.2
Flask-Jsonpify==1.1
gunicorn==19.9.0
html2text==2018.1.9
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nltk==3.4
pycodestyle==2.5.0
pymongo==3.7.2
python-dateutil==2.8.0
singledispatch==3.4.0.3
six==1.12.0
Werkzeug==0.15.2